services/beit3-image-text-retrieval/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic-settings==2.0.3


# Models
torch==2.0.0
torchvision==0.15.1
torchscale==0.2.0
ftfy==6.1.1
regex
tqdm==4.66.1
transformers==4.33.1
timm==0.4.12
sentencepiece==0.1.99

# Vector Database
faiss-cpu

# Project settings, for development only
# pre-commit